refactor: Centralize useGetBudgetSettings hook and types
