+++ js/src/domains/settings/budget/types.ts
@@ -1,11 +1,3 @@
-export type BudgetSettings = {
-    id: number;
-    local_currency: string;
-    exchange_rate: string;
-    inflation_rate: string;
-    buffer: string;
-};
-
 export type BudgetSettingsPayload = {
     id: number;
     local_currency: string;
