--- js/src/domains/settings/interventions/types/budgetSettings.ts
@@ -1,6 +1,0 @@
-export type BudgetSettings = {
-    id: number;
-    currency: string;
-    exchange_rate: number;
-    inflation_rate: number;
-};
